hr_onboarding_knowledge_assistant/pyproject.toml
[project]
name = "hr-onboarding-knowledge-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.6",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "streamlit>=1.46.1",
]
